+++ packages/quarks.examples/emitterShapeDemo.js
@@ -27,6 +27,7 @@ import {
 import {TextGeometry} from 'three/examples/jsm/geometries/TextGeometry.js';
 import {Demo} from './demo.js';
 import {FontLoader} from 'three/examples/jsm/loaders/FontLoader.js';
+import helvetikerRegularUrl from 'three/examples/fonts/helvetiker_regular.typeface.json?url';
 
 export class EmitterShapeDemo extends Demo {
     name = 'Different Emitter Shapes';
@@ -75,7 +76,7 @@ export class EmitterShapeDemo extends Demo {
 
         const loader = new FontLoader();
 
-        loader.load( 'js/examples/fonts/helvetiker_regular.typeface.json', ( font ) => {
+        loader.load( helvetikerRegularUrl, ( font ) => {
 
             const fontConfig = {
                 font: font,